setsockopt and getsockopt use int as others functions
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@17913 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/legacy/network/sys/socket.h
@@ -277,8 +277,8 @@ ssize_t recv(int, void *, size_t, int);
 ssize_t sendto(int, const void *, size_t, int, const struct sockaddr *, int);
 ssize_t recvfrom(int, void *, size_t, int, struct sockaddr *, int *);
 
-int     setsockopt(int, int, int, const void *, size_t);
-int     getsockopt(int, int, int, void *, size_t *);
+int     setsockopt(int, int, int, const void *, int);
+int     getsockopt(int, int, int, void *, int *);
 int     getpeername(int, struct sockaddr *, int *);
 int     getsockname(int, struct sockaddr *, int *);
 #endif /* _KERNEL_MODE */
